Added credits feature for media elements
Added the credit section for cover images with css. And created a new style for crediting images and other media.

The crediting covers works by reading the YAML markup variable `creditCover` and `creditText` in the page

Crediting in images and other media works by adding the style to the credit text on the respective page.